Refactoring of Pattern class.
Major refactoring concerning the "_size" data member of the Pattern
class. Before this member described the size of one packed element of an
element using a compile-time-evaluatable Pattern. For non-compile-time-
evaluatable Patterns it described the size of the entire packed range.
This duality is obviously a bad design. To fix this the "_size" member
was removed and replaced with functions that return the packed size of
one element or the size of a packed size of a given range. Since _size
was the last data member of Pattern, the constructors could be
significantly simplified.