Major interface change: Introduce Pattern view.
In this commit we introduce the new Pattern view. This changes the
interface for sending data with a pattern from:

comm.send(Destination(0), Tag(0), pattern, data);

to:

comm.send(Destination(0), Tag(0), data | pattern_view(pattern));

This eliminates the need for the Pattern parameter in the send function
and combines everything into the input range parameter.
Alongside this long overdue changes to the send() and recv() functions
in the communicator have been undertaken. This in combination with the
interface changes makes the code more user- and deverloper friendly.